+++ src/Resolution.ts
@@ -211,15 +211,37 @@ export default class Resolution {
    * @throws [[ResolutionError]] if address is not found
    * @returns A promise that resolves in an address
    */
-  async addr(domain: string, currrencyTicker: string): Promise<string> {
+  async addr(domain: string, ticker: string): Promise<string> {
     return await this.record(
       domain,
-      `crypto.${currrencyTicker.toUpperCase()}.address`,
+      `crypto.${ticker.toUpperCase()}.address`,
     );
   }
 
   /**
+   * Read multi-chain currency address if exists
+   * @async
+   * @param domain - domain name to be resolved
+   * @param ticker - currency ticker (USDT, FTM, etc.) 
+   * @param chain - chain version, usually means blockchain ( ERC20, BEP2, OMNI, etc. )
+   * @throws [[ResolutionError]] if address is not found
+   * @returns A promise that resolves in an adress
+   */
+  async multiChainAddr(domain: string, ticker: string, chain: string): Promise<string> {
+    domain = this.prepareDomain(domain);
+    const method = this.getNamingMethodOrThrow(domain);
+    if (method.name === NamingServiceName.ENS) {
+      throw new ResolutionError(ResolutionErrorCode.UnsupportedMethod,
+        { methodName: NamingServiceName.ENS, domain });
+    }
+
+    const recordKey = `crypto.${ticker.toUpperCase()}.version.${chain.toUpperCase()}.address`;
+    return await method.record(domain, recordKey);
+  }
+
+  /**
    * Resolves given domain name to a specific USDT chain address if exists
+   * @deprecated 
    * @async
    * @param domain - domain name to be resolved
    * @param version - chain version to look for such as ERC20, TRON, EOS, OMNI